Adding comments and logging
Further comments could be added into every method, but for now javadoc
for every method is set.
Logging is setup to allow TRACE logging if needed, but stores INFO and
WARN to separate files.
WARN is also set to show up in console, usually the only level you care
about on the console.